requirements.txt
accelerate
datasets
faiss-cpu
numpy
pandas
pyarrow
sentence-transformers
streamlit
torch
tqdm
transformers
